# add some speed-up tricks